+++ src/main/java/io/vntr/spar/SparMiddleware.java
@@ -168,12 +168,12 @@ public class SparMiddleware implements IMiddleware, IMiddlewareAnalyzer {
 
     @Override
     public Integer getNumberOfPartitions() {
-        return (int) manager.getAllPartitionIds().size();
+        return manager.getAllPartitionIds().size();
     }
 
     @Override
     public Integer getNumberOfUsers() {
-        return (int) manager.getNumUsers();
+        return manager.getNumUsers();
     }
 
     @Override
